getDivide

+++ src/main/java/Cal.java
@@ -1,5 +1,5 @@
 public class Cal {
-    public double getDivide(int a, int b) {
+    public int getDivide(int a, int b) {
         return a / b;
     }
 }
